Implemented a simple addon mechanism
If asked to do so, mkarchiso simply copies a
directory full of addons to the iso root.

On boot, after union-mounting /real_root, the
archiso hook will look for and source an addon
config file. This file is a plain old bash
script, which makes it quite flexible. The addon
config should be written to take care of any
mounting that needs to be done, an example of
typical tasks is also included.

Signed-off-by: Simo Leone <simo@archlinux.org>