Rename ramdisk_size parameter to tmpfs_size to avoid conflict
Rename ramdisk_size parameter to tmpfs_size to avoid conflict with
the kernel parameter that is for old ramdisk based blockdevs (/dev/ramN)
and confusion to users.

The kernel parameter ramdisk_size=N is in "N * k size", while tmpfs
size=N is in "N * [ k | m | g ] byte size" or "in N% of RAM",
like is used by default on archiso.

This parameter is used only for tmpfs in size option.

Signed-off-by: Gerardo Exequiel Pozzi <vmlinuz386@yahoo.com.ar>